+++ CHANGELOG.md
@@ -3,6 +3,5 @@
 ## 0.1.0.0 -- YYYY-MM-DD
 
 * Initial release
-* Skeleton structure with `S2nTlsSys` record type
-* Support for linked bindings via `linked` flag
-* Support for dynamic loading via `dynamic` flag
+* Skeleton structure with `S2nTlsFfi` record type
+* Support for linked and dynamic loading via `Library` type